Clarify README and add .gitignore
Three file changes are included in this commit to make the example
more accessible to new DevOps engineers.

1. A .gitignore file is added to the repo so that when the executable
is run locally the SQLite database that gets created is ignored in
subsequent pushes.
2. README.md is upated with a longer explanation of how to run the
container locally before building the AWS infrastructure.  The
instructions in the README as it originally existed did not publish
port 8080 so if an unfamiliar Docker user ran the container they
could not connect to port 8080 because it was connected to the host
port 8080.  This updates the READM to clarify that point and explain
a few of the Docker flags that make it work.
3. Small change to .circleci.yml to get back to a generic filepath for
after I had updated it to my specific repo.  This ensures that if this
pull request is merged there is still clear guidance for users on how
rename the filepath to their Github repo to trigger CircleCI.